Refactor everything and organize to packages
