+++ data/tasks.txt
@@ -3,6 +3,5 @@ D#true#poopy#2034-02-12T19:28
 E#false#poopi#2034-02-12T19:28#2034-03-12T19:29
 T#true#pooa
 T#true#pooper
-T#false#poopydoopy
 T#false#aks
 T#false#aks
