--- src/main/java/msrainy/command/List.java
@@ -1,4 +1,0 @@
-package main.java.command;
-
-public class List {
-}
